debugged item spawn
removed some debugging lines.
changed the values of the points that the system looks for. changed the if for the spawn points so they it changes the string proprly. commented out the line that changes the position of the items.
items now actually sapwning at the points that there suppused to still some stuff that needs to be changed like scaling trhe size of the gear and weaponm aswell as haveing it be able to spawn multiple weapons (jeff i have i way to do this)